moving payments page to correct folder, moving payments under pages
