+++ pages/payments/sections/AdvancedFeaturesSection.tsx
