+++ pages/payments/sections/DeveloperSpotlightSection.tsx
