+++ pages/payments/sections/EmbeddedPaymentsSection.tsx
