+++ pages/payments/sections/FlexibleIntegrationSection.tsx
