+++ pages/payments/sections/HeroSection.tsx
