+++ pages/payments/sections/PartnerLogosSection.tsx
