+++ pages/payments/sections/StablecoinsSection.tsx
