+++ pages/payments/sections/StayConnectedSection.tsx
