+++ pages/payments/sections/WhyChooseSection.tsx
